OOP
It is a text-based RPG where you play as a character who must fight monsters in order to survive. You start the game at level one and with 10 health points. As you kill monsters, you gain experience and level up. Each level increases your character's attack power.

To fight a monster, you must enter the command "F" or "f". You can either attack the monster or try to escape. If you try to escape, you have a 50% chance of success. If you fail to escape, the monster will attack you.

The goal of the game is to defeat as many monsters as possible and increase your level. The game ends when you die or defeat all monsters.

Tips for the game:

Try not to run away from monsters too often, otherwise you won't be able to level up.
Use your attack power wisely. If you are low on health, it is better to retreat and regain your health than to attack the monster and die.
Don't forget to collect the gold that monsters drop. Gold can be used to purchase items that will help you in the game.